requirements.txt
pvporcupine>=3.0.0
pyaudio>=0.2.14
mss>=9.0.0
sounddevice>=0.4.6
scipy>=1.11.0
openai>=1.0.0
google-generativeai>=0.5.0
pygame>=2.5.0
python-dotenv>=1.0.0
